backend/requirements.txt
# Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
email-validator==2.3.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.6
numpy==1.26.3
scipy==1.12.0
pytest==7.4.4
httpx==0.26.0
google-generativeai
python-dotenv
python-multipart
alembic
beautifulsoup4
requests
